requirements.txt
docx2txt==0.9
fitz==0.0.1.dev2
mysql_connector_repackaged==0.3.1
numpy==2.4.4
ollama==0.6.1
Pillow==12.2.0
python_pptx==1.0.2
sentence_transformers==5.1.2
watchdog==6.0.0
python-magic-bin
